deceleration

+++ src/asteroids/constants.h
@@ -6,7 +6,7 @@ extern const int WIN_HEIGHT;
 extern const float SPLASH_DURATION;
 
 extern const float MAX_VELOCITY; //pixels per second
-extern const int SPEED;
+extern const float SPEED;
 
 extern const int BULLET_SPEED;
 extern const float FIRE_RATE; //bullets per minute
